GuardedCommands and Commands
I have done all the work for the GuardedCommands and Commands in the Parser, Lexer and AST.

It is all commented out since it is not running because of the definition of the commands and guardcommands in the AST-file. It runs with them commented out.

Solution is on its way.